made getPitchFromDerivative abstract and added computePitchWeight
